persist participants

+++ src/main/scala/service/ParticipantService.scala
@@ -14,11 +14,12 @@ object ParticipantService {
 
   def apply(participantRepository: ParticipantRepository): ParticipantService =
     new ParticipantService {
+
       def register(request: RegisterParticipantRequest): IO[Either[String, RegisterParticipantResponse]] = {
         val validationResult: Either[NonEmptyList[String], Participant] = ParticipantValidator.validateParticipant(request).toEither
         validationResult match
           case Left(errors)       => IO.pure(Left(s"The following validation(s) failed: ${errors.toList.mkString(",")}."))
-          case Right(participant) => participantRepository.register(participant).map(id => Right(RegisterParticipantResponse(id)))
+          case Right(participant) => participantRepository.register(participant).map(id => Right(RegisterParticipantResponse(id.toInt)))
       }
     }
 }
